+++ KadoPanel/src/config.py
@@ -1,14 +1,17 @@
 # -*- coding: utf-8 -*-
 
 PANEL_NAME = "Kado Panel"
-VERSION = "0.1.0 Alpha"
+VERSION = "0.2.0 Alpha"
 AUTHOR = "Captain Essam"
 
 LOG_FILE = "/tmp/KadoPanel.log"
 CACHE_DIR = "/tmp/kadopanel/"
 PROFILE_DIR = "/etc/kadopanel/profiles/"
 BACKUP_DIR = "/etc/kadopanel/backups/"
+REPORT_FILE = "/tmp/KadoPanel-SystemReport.txt"
 
 SUPPORTED_IMAGES = ["OpenBH", "OpenViX", "OpenATV", "Egami", "OpenPLi", "PurE2"]
 PRIMARY_TEST_IMAGE = "Open Black Hole 5.6.008"
 NEOBOOT_TARGET = "9.65"
+
+GITHUB_REPOSITORY = "https://github.com/kado1957/KadoPanel"